Adapt for changes to Hash_Get_Size.
The return value has been widened to size_t, so cast where necessary.

+++ core/Lucy/Index/Snapshot.c
@@ -88,7 +88,7 @@ Snapshot_List_IMP(Snapshot *self) {
 uint32_t
 Snapshot_Num_Entries_IMP(Snapshot *self) {
     SnapshotIVARS *const ivars = Snapshot_IVARS(self);
-    return Hash_Get_Size(ivars->entries);
+    return (uint32_t)Hash_Get_Size(ivars->entries);
 }
 
 void
